Only remove entity if it still exists.

+++ src/Contao/Doctrine/ORM/DataContainer/General/EntityData.php
@@ -380,8 +380,10 @@ class EntityData implements DriverInterface
 				->find($item);
 		}
 
-		$entityManager->remove($entity);
-		$entityManager->flush($entity);
+		if ($entity) {
+			$entityManager->remove($entity);
+			$entityManager->flush($entity);
+		}
 	}
 
 	/**
